src/requirements.txt
environ==1.0
whitenoise==6.2.0
Django==4.0.6
django-environ==0.9.0
django-filter==22.1
django-localflavor==3.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.2.0
djangorestframework-dataclasses==1.1.1
drf-spectacular==0.22.1
requests==2.28.1
pytest==7.1.2
pytest-django==4.5.2
pytest-cov==3.0.0
pytest-vcr==1.0.2
pytest-socket==0.5.1
psycopg2-binary==2.9.3
PyJWT==2.4.0
Faker==13.15.0
gunicorn==20.1.0
isort==5.10.1
mypy==0.961
black==22.3.0
model_bakery==1.6.0
tzdata==2022.1
types-requests==2.28.0
